+++ packages/assignment-4/src/basic/constant.js
@@ -1,0 +1,5 @@
+export const OPTIONS = [
+  { id: "p1", name: "상품1", cost: 10000 },
+  { id: "p2", name: "상품2", cost: 20000 },
+  { id: "p3", name: "상품3", cost: 30000 },
+];
